archifarm-core/experiment-management-cli/pyproject.toml
[tool.poetry]
name = "experiment-management-cli"
version = "0.1.0"
description = "A CLI used to generate and submit experiment requests to the testbed"
authors = ["CREST <crest@adelaide.edu.au>"]
packages = [{include = "src", from = "."}]
package-mode = true

[tool.poetry.scripts]
experiment-cli = "main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "0.12.5"
httpx = "0.27.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
rich = "^13.8.1"
tomli = "^2.0.1"
loguru = "^0.7.2"
toml = "^0.10.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ruff = "*"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "UP", # pyupgrade
    "I",  # isort
]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "A",
    "B",
    "C",
    "D",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ANN",
    "ARG",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py39"

[tool.ruff.lint]

select = [
    "E", # pycodestyle errors (settings from FastAPI, thanks, @tiangolo!)
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501",
    "C901", # too complex
]

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
known-first-party = []

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 72

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"